requirements.txt
python-dotenv
faiss-cpu
langchain-groq
langchain-huggingface
langchain-community
langchain-core
langchain-text-splitters
streamlit